rename: panel class 생성
