+++ src/main/java/org/teamseven/tetris/score/test2.csv
@@ -1,0 +1,7 @@
+"AAA","100","2022-04-20","b7a6e019-0707-4510-b0bb-a166df917b68"
+"AAA","200","2022-04-20","9fe7365d-022e-44b2-8895-5a903b5bca6a"
+"AAB","100","2022-04-20","9b374c4d-0641-4e59-ae2e-7a5db8569cb6"
+"AAA","100","2022-04-20","42861620-a6d3-43aa-8dae-f55a06907a64"
+"AAA","200","2022-04-20","c51ae435-f60b-4407-8f66-ac83cf1a7bd1"
+"AAB","100","2022-04-20","a60000ab-38c6-4335-8496-8843b047ab67"
+"AAA","100","2022-04-20","1aca5af9-0a04-41c7-a80b-26b8c0366050"
