+++ src/main/java/org/teamseven/tetris/ui/ScoreBoardTabbedPane.java
@@ -20,7 +20,7 @@ import java.util.Map;
 
 public class ScoreBoardTabbedPane extends JTabbedPane implements  IDesign {
     private JLabel lTime;
-
+    private ScoreBoardPanelTab noItem, item;
     private int[] preferredResolution;  // frame resolution - frame top border
     private JLabel title;
 
@@ -38,7 +38,8 @@ public class ScoreBoardTabbedPane extends JTabbedPane implements  IDesign {
     private GridBagConstraints gridBagConstraints;
     private GridBagLayout gridBagLayout;
     private ScoreHandler handler = new ScoreHandler();
-    private String fileName = "test.csv";
+    private String noItemCvs = "test.csv";
+    private String itemCvs = "test2.csv";
 
 
     public ScoreBoardTabbedPane(int[] preferredResolution) {
@@ -54,271 +55,24 @@ public class ScoreBoardTabbedPane extends JTabbedPane implements  IDesign {
     @Override
     public void setComp() {
 
-        gridBagConstraints = new GridBagConstraints();
-        gridBagLayout = new GridBagLayout();
-
-        noItemScorePanelTab = new JPanel();
-        itemScorePanelTab = new JPanel();
-
-        titlePanel = new JPanel();
-        scorePanel1st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel2st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel3st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel4st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel5st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel6st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel7st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel8st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel9st = new JPanel(new GridLayout(0,4,0,0));
-        scorePanel10st = new JPanel(new GridLayout(0,4,0,0));
-
-        scorePanels = new JPanel[]{scorePanel1st, scorePanel2st, scorePanel3st, scorePanel4st, scorePanel5st,
-                scorePanel6st, scorePanel7st, scorePanel8st, scorePanel9st, scorePanel10st};
-
-        line1st = new JPanel();
-        line2st = new JPanel();
-        line3st = new JPanel();
-        line4st = new JPanel();
-        line5st = new JPanel();
-        line6st = new JPanel();
-        line7st = new JPanel();
-        line8st = new JPanel();
-        line9st = new JPanel();
-        line10st = new JPanel();
-
-        buttonPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 130, 0));
-
-        title = new JLabel("SCORE");
-
-        leftButton = new CustomButton();
-        rightButton = new CustomButton();
-
-
+        noItem = new ScoreBoardPanelTab(preferredResolution,noItemCvs);
+        item = new ScoreBoardPanelTab(preferredResolution,itemCvs);
 
     }
 
     @Override
     public void setDesign() {
-
-        gridBagConstraints.fill = GridBagConstraints.BOTH;
-
-        noItemScorePanelTab.setLayout(gridBagLayout);
-
-        titlePanel.add(title);
-        titlePanel.setPreferredSize(new Dimension(preferredResolution[1] * 8 / 130,preferredResolution[1] * 8 / 130));
-        title.setFont(new Font("ss",Font.CENTER_BASELINE,preferredResolution[1] * 10 / 130));
-        title.setForeground(Color.orange);
-
-        draw(pageNum);
-
-        gridBagConstraints.weightx = 1.0;
-        gridBagConstraints.weighty = 4.0;
-
-        gridBagConstraints.insets = new Insets(preferredResolution[1] * 4 / 130, 0, 0, 0);
-        make(noItemScorePanelTab, titlePanel,1,0,1,1);
-        gridBagConstraints.weighty = 2.0;
-        gridBagConstraints.insets = new Insets(preferredResolution[1] * 4 / 130, preferredResolution[0] / 6, 0, preferredResolution[0] / 8);
-
-        make(noItemScorePanelTab, scorePanel1st,1,1,1,1);
-
-        gridBagConstraints.insets = new Insets(0, preferredResolution[0] / 6, 0, preferredResolution[0] / 8);
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line1st,1,2,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel2st,1,3,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line2st,1,4,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel3st,1,5,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line3st,1,6,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel4st,1,7,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line4st,1,8,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel5st,1,9,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line5st,1,10,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel6st,1,11,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line6st,1,12,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel7st,1,13,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line7st,1,14,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel8st,1,15,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line8st,1,16,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel9st,1,17,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line9st,1,18,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        make(noItemScorePanelTab, scorePanel10st,1,19,1,1);
-
-        gridBagConstraints.weighty = 0.1;
-        make(noItemScorePanelTab, line10st,1,20,1,1);
-        gridBagConstraints.weighty = 2.0;
-
-        gridBagConstraints.insets = new Insets(0, 0, 0, 0);
-        make(noItemScorePanelTab, buttonPanel,1,21,1,1);
-
-        noItemScorePanelTab.setBackground(Color.white);
-
-        this.addTab("noitem",noItemScorePanelTab);
-        this.addTab("item",itemScorePanelTab);
-
-
-        leftButton.setFont(new Font("ss",Font.BOLD,preferredResolution[1] * 8 / 130));
-        leftButton.setForeground(Color.gray);
-        rightButton.setFont(new Font("ss",Font.BOLD,preferredResolution[1] * 8 / 130));
-        rightButton.setForeground(Color.gray);
-
-
-        rightButton.requestFocus();
-        leftButton.setText("<");
-        rightButton.setText(">");
-        buttonPanel.add(leftButton);
-        buttonPanel.add(rightButton);
-
-        noItemScorePanelTab.setFocusCycleRoot(true);
-
-
-
+        this.addTab("noitem",noItem);
+        this.addTab("item",item);
 
 
     }
 
     @Override
     public void setAction() {
-        leftButton.removeFocusListener(leftButton.getFocusListeners()[1]);
-        leftButton.removeFocusListener(leftButton.getFocusListeners()[0]);
 
-        rightButton.removeFocusListener(rightButton.getFocusListeners()[1]);
-        rightButton.removeFocusListener(rightButton.getFocusListeners()[0]);
+        this.setFocusable(false);
 
-        leftButton.addFocusListener(new FocusListener() {
-            @Override
-            public void focusGained(FocusEvent e) {
-                leftButton.setForeground(Color.darkGray);
-            }
-
-            @Override
-            public void focusLost(FocusEvent e) {
-                leftButton.setForeground(Color.gray);
-
-            }
-        });
-        rightButton.addFocusListener(new FocusListener() {
-            @Override
-            public void focusGained(FocusEvent e) {
-                rightButton.setForeground(Color.darkGray);
-            }
-
-            @Override
-            public void focusLost(FocusEvent e) {
-                rightButton.setForeground(Color.gray);
-
-            }
-        });
-
-
-                leftButton.addActionListener(e -> {
-                    if (pageNum > 0) {
-                        pageNum--;
-                        clear();
-                        draw(pageNum);
-                        this.revalidate();
-                        this.repaint();
-                    }
-                });
-
-        rightButton.addActionListener(e -> {
-            if(pageNum < totalPageNum) {
-                pageNum++;
-                clear();
-                draw(pageNum);
-
-            }
-        });
-
-
-    }
-
-    public void make(JPanel p, JComponent c, int x, int y, int w, int h) {
-        gridBagConstraints.gridx = x;
-        gridBagConstraints.gridy = y;
-        gridBagConstraints.gridwidth = w;
-        gridBagConstraints.gridheight = h;
-        p.add(c, gridBagConstraints);
-
-        // for test
-        c.setBackground(Color.white);
-        // c.setBackground(Color.getHSBColor((y+1)  /9f + (x+1) /9f,0.75f,0.95f));
-        //  gridBagConstraints.insets = new Insets(15,10,15,10);
-
-    }
-
-    public  void draw(int pageNum){
-        List<Score> scores = handler.getScores(fileName);
-        totalPageNum = scores.size() % 10;
-        for(int i = 0; i<10; i++){
-            if(scores.size() == pageNum*10 + i){break;};
-            JLabel tempLabel = new JLabel();
-            tempLabel.setFont(new Font("ss",Font.BOLD,preferredResolution[1] * 4 / 130));
-            tempLabel.setForeground(Color.gray);
-            tempLabel.setText(String.valueOf(pageNum*10+i+1));
-            scorePanels[i].add(tempLabel);
-
-            JLabel tempLabel2 = new JLabel();
-            tempLabel2.setFont(new Font("ss",Font.BOLD,preferredResolution[1] * 4 / 130));
-            tempLabel2.setForeground(Color.gray);
-            tempLabel2.setText(scores.get(pageNum*10+i).getName());
-            scorePanels[i].add(tempLabel2);
-
-            JLabel tempLabel3 = new JLabel();
-            tempLabel3.setFont(new Font("ss",Font.BOLD,preferredResolution[1] * 4 / 130));
-            tempLabel3.setForeground(Color.gray);
-            tempLabel3.setText(scores.get(pageNum*10+i).getScore());
-            scorePanels[i].add(tempLabel3);
-
-            JLabel tempLabel4 = new JLabel();
-            tempLabel4.setFont(new Font("ss",Font.BOLD,preferredResolution[1] * 4 / 130));
-            tempLabel4.setForeground(Color.gray);
-            tempLabel4.setText(scores.get(pageNum*10+i).getDate());
-            scorePanels[i].add(tempLabel4);
-
-            scorePanels[i].setPreferredSize(new Dimension(preferredResolution[1] * 8 / 130,preferredResolution[1] * 3 / 130));
-            this.revalidate();
-            this.repaint();
-        }
-    }
-
-    public  void clear(){
-        for(int i = 0; i<10; i++) {
-            scorePanels[i].removeAll();
-
-        }
 
     }
 }
